feat(task) : Manage to create, patch and delete task inside BlockUserView
When PATCH to BlockUserView view, this class will now check if the Body contains Task data.
If there are no Tasks in the Body, all Tasks associated with the Block will be DELETE. If there is data containing a Task ID, the Task will be PATCH and if there is none, a new Task will be Created.